Move set_typing_action to confs and import to all
