+++ meldebot/confs.py
@@ -75,3 +75,14 @@ def get_giphy_api_key():
 
 def get_debug_enabled():
     return config.get('DEVELOPMENT', 'debug')
+
+def send_typing_action(func):
+    """Sends typing action while processing func command."""
+
+    @wraps(func)
+    def command_func(update, context, *args, **kwargs):
+        context.bot.send_chat_action(chat_id=update.effective_message.chat_id,
+                                     action=telegram.ChatAction.TYPING)
+        return func(update, context,  *args, **kwargs)
+
+    return command_func
