+++ meldebot/mel.py
@@ -17,18 +17,7 @@ from functools import wraps
 
 # Self imports
 from meldebot.confs import get_giphy_api_key, get_debug_enabled
-
-
-def send_typing_action(func):
-    """Sends typing action while processing func command."""
-
-    @wraps(func)
-    def command_func(update, context, *args, **kwargs):
-        context.bot.send_chat_action(chat_id=update.effective_message.chat_id,
-                                     action=telegram.ChatAction.TYPING)
-        return func(update, context,  *args, **kwargs)
-
-    return command_func
+from meldebot.confs import send_typing_action
 
 
 def get_gif_url(params):
