+++ meldebot/tuquiets.py
@@ -1,6 +1,8 @@
 from telegram.ext import CommandHandler
 from random import randint
 
+from meldebot.confs import send_typing_action
+
 def get_random_name():
     name = [
      "Jaumayer",
@@ -17,6 +19,7 @@ def get_random_name():
     name_num = randint(0, len(name) -1)
     return name[name_num]
 
+@send_typing_action
 def get_tuquiets_text(update, context):
     _name = get_random_name()
     update.message.reply_text(_name)
